Don't raise in Helpers#current_user
The normal api for this helper method is to return nil if there is
no current user. Controllers using this method (or a derived one like
signed_in?) won't want to take care of cases where the user is either
not signed in at all versus using an outdated token etc. They'll just
want to redirect (or whatever) if there's no current user.